Revert "Add request parsing"
