+++ CMakeLists.txt
@@ -8,8 +8,8 @@ project(${TARGET}
 
 add_executable(${TARGET}
     src/main.cpp
-    src/HttpServer.cpp
-    include/HttpServer.h
+    src/ThreadPool.cpp
+    include/ThreadPool.h
     include/AtomicQueue.h
     include/constants.h)
